+++ Vizsgaremek/Models/DailyTarget.cs
@@ -5,10 +5,10 @@
         public int Id { get; set; }
         public string UserId { get; set; }
         public User User { get; set; }
-        public double Calories { get; set; }
-        public double Protein { get; set; }
-        public double Carbohydrates { get; set; }
-        public double Fats { get; set; }
+        public decimal Calories { get; set; }
+        public decimal Protein { get; set; }
+        public decimal Carbohydrates { get; set; }
+        public decimal Fats { get; set; }
         public DateOnly Date { get; set; }
 }
 }
