+++ Vizsgaremek/Models/Recipe.cs
@@ -9,10 +9,10 @@
         public int PreparationTime { get; set; }
         public int CookingTime { get; set; }
         public int Portions { get; set; }
-        public double Calories { get; set; }
-        public double Protein { get; set; }
-        public double Carbohydrates { get; set; }
-        public double Fats { get; set; }
+        public decimal Calories { get; set; }
+        public decimal Protein { get; set; }
+        public decimal Carbohydrates { get; set; }
+        public decimal Fats { get; set; }
         public bool IsCommunity { get; set; }
         public ICollection<RecipeIngredient> RecipeIngredients { get; set; }
     }
