+++ Vizsgaremek/Models/UserActivity.cs
@@ -9,9 +9,9 @@ namespace Vizsgaremek.Models
         public User User { get; set; }
         public int ActivityId { get; set; }
         public Activity Activity { get; set; }
-        public double Duration { get; set; }
+        public decimal Duration { get; set; }
         public DateTime ActivityDate { get; set; }
         [NotMapped]
-        public double CaloriesBurned => (Activity.CaloriesBurnedPerHour * Duration) / 60.0;
+        public decimal CaloriesBurned => (Activity.CaloriesBurnedPerHour * Duration) / 60m;
     }
 }
