+++ Vizsgaremek/Models/UserAttributes.cs
@@ -7,12 +7,12 @@ namespace Vizsgaremek.Models
         public int Id { get; set; }
         public string UserId { get; set; }
         public User User { get; set; }
-        public double Weight { get; set; } 
-        public double Height { get; set; }
+        public decimal Weight { get; set; } 
+        public decimal Height { get; set; }
         public DateTime MeasuredAt { get; set; }
 
         [NotMapped]
-        public double Bmi => Math.Round(Weight / ((Height / 100) * (Height / 100)), 2);
+        public decimal Bmi => Math.Round(Weight / ((Height / 100m) * (Height / 100m)), 2);
 
     }
 }
